refactor: extract QueryClassifier from MainContentCoordinator for MCP reuse
